+++ eagle-core/eagle-application-management/eagle-application-service/src/main/scala/org/apache/eagle/stream/application/model/EntityConversion.scala
