+++ eagle-core/eagle-application-management/eagle-stream-application-manager/src/test/resources/application.conf
