+++ eagle-core/eagle-application-management/eagle-stream-application-manager/src/test/resources/log4j.properties
