--- eagle-core/eagle-data-process/eagle-stream-application-manager/src/main/scala/org/apache/eagle/stream/application/model/TopologyOperationModel.scala
@@ -1,27 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *
- */
-
-package org.apache.eagle.stream.application.model
-
-import org.apache.eagle.stream.application.entity.TopologyOperationEntity
-
-
-case class TopologyOperationModel(site: String, application: String, topology: String, uuid: String, operation: String, status: String, lastModifiedDate: Long) extends EntityConversion[TopologyOperationEntity] {
-  override def toEntity: TopologyOperationEntity = TopologyOperationEntity.fromModel(this)
-  override def fromEntity(entity: TopologyOperationEntity) = TopologyOperationEntity.toModel(entity)
-}
