+++ eagle-core/eagle-data-process/pom.xml
@@ -32,6 +32,5 @@
     	<module>eagle-job-common</module>
         <module>eagle-stream-process-api</module>
         <module>eagle-stream-pipeline</module>
-        <module>eagle-stream-application-manager</module>
     </modules>
 </project>
