+++ eagle-webservice/pom.xml
@@ -146,11 +146,12 @@
 			<artifactId>eagle-security-hdfs-web</artifactId>
 			<version>${project.version}</version>
 		</dependency>
+		<!--
 		<dependency>
 			<groupId>org.apache.eagle</groupId>
 			<artifactId>eagle-stream-application-manager</artifactId>
 			<version>${project.version}</version>
-		</dependency>
+		</dependency> -->
         <dependency>
             <groupId>org.apache.eagle</groupId>
             <artifactId>eagle-topology-assembly</artifactId>
